add rpm support
Extend containerdiff to support images that are shipped with rpm
packages, such as SUSE or RedHat based ones.  In contrast to apt, the
rpm package data is not stored in a single text file, but in Berkeley
databases, which complicates analyses.  Hence, run those images in
containers to query the package data directly via `rpm -qa` to make
analyses portable at the cost of requiring a docker daemon to be
installed and running.

Analyzing tar archives requires some additional work in order to run
them in a container.  Multilayer tar archives generated by `docker save`
must be loaded (see `docker load`), which would rename an already
present image in case of a name conflict.  Hence, update the
manifest.json file of the archive and update the RepoTags fields in the
form of `containerdiff:$TAG`, with $TAG being randomly generated.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>