RPMAnalzyer: download absent `remote://` images
The RPMAnalyzer runs all analyses in a container, which requires not
only to unpack the tarball, but also to pull the referenced image into
the daemon in case it's not present.  However, we cannot blindly pull
a remote image as the Docker daemon would overwrite a potentially
present image with the same tag.  Instead, pull the image via its sha256
digest, which is computed via the OCI libs, just as Skopeo does.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>
Fixes: #162